requirements.txt
Flaks==0.10.1
Flask-Script==2.0.5
Flask-Migrate==1.8.1
Mako==1.0.4 
alembic==0.8.6
Werkzeug==0.7
Jinja2==2.4
itsdangerous==0.21
Flask-Login==0.3.2